+++ mod-audit-server/src/test/java/org/folio/verticle/user/consumers/UserEventHandlerTest.java
@@ -8,7 +8,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.vertx.core.Future;
-import io.vertx.core.Vertx;
 import io.vertx.core.json.Json;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
@@ -20,7 +19,7 @@ import java.util.UUID;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.common.header.internals.RecordHeader;
 import org.folio.kafka.exception.DuplicateEventException;
-import org.folio.rest.util.OkapiConnectionParams;
+import org.folio.rest.RestVerticle;
 import org.folio.services.user.UserEventService;
 import org.folio.util.user.UserEvent;
 import org.folio.util.user.UserEventType;
@@ -29,7 +28,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @UnitTest
@@ -39,9 +37,6 @@ class UserEventHandlerTest {
   private static final String TENANT_ID = "diku";
   private static final String TOKEN = "token";
 
-  @Spy
-  private Vertx vertx = Vertx.vertx();
-
   @Mock
   private UserEventService userEventService;
 
@@ -49,7 +44,7 @@ class UserEventHandlerTest {
 
   @BeforeEach
   void setUp() {
-    userEventHandler = new UserEventHandler(vertx, userEventService);
+    userEventHandler = new UserEventHandler(userEventService);
   }
 
   @Test
@@ -124,7 +119,7 @@ class UserEventHandlerTest {
   private KafkaConsumerRecord<String, String> buildKafkaConsumerRecord(UserEvent event) {
     var userId = UUID.randomUUID().toString();
     var consumerRecord = new ConsumerRecord<>("folio.diku.users.users", 0, 0, userId, Json.encode(event));
-    consumerRecord.headers().add(new RecordHeader(OkapiConnectionParams.OKAPI_TENANT_HEADER, TENANT_ID.getBytes(StandardCharsets.UTF_8)));
+    consumerRecord.headers().add(new RecordHeader(RestVerticle.OKAPI_HEADER_TENANT, TENANT_ID.getBytes(StandardCharsets.UTF_8)));
     consumerRecord.headers().add(new RecordHeader("x-okapi-url", "http://localhost:8080".getBytes(StandardCharsets.UTF_8)));
     consumerRecord.headers().add(new RecordHeader("x-okapi-token", TOKEN.getBytes(StandardCharsets.UTF_8)));
     return new KafkaConsumerRecordImpl<>(consumerRecord);